UPSTREAM: drm: bridge: ti-sn65dsi83: Move LVDS format selection to .mode_set()
The LVDS format is selected based on the bus format reported by the
connector. This is currently done in .mode_fixup(), but that's not the
right place, as the format should be selected when setting the mode.
Move it to .mode_set().

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Acked-by: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Robert Foss <robert.foss@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210621125518.13715-2-laurent.pinchart@ideasonboard.com
(cherry picked from commit b966d857c57acb5cd2b12e5e6e466c980d90d362)

BUG=b:228272255
BUG=b:191366271
TEST=for arch in amd64-generic arm64-generic; do \
	USE=buildtest emerge-"${arch}" chromeos-kernel-5_10; \
     done

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I5c70316b9bf249d0d9918751a9c51f3e142b6afd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3574403
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>