CHROMIUM: mm: multigenerational lru: scan kvm mmu pages
Reland of:
https://crrev.com/c/2987928

For each kvm instance, we add a list of its mmu pages. And each kvm
mmu page is reference counted and will not be freed before an rcu
grace period has expired. This allows us to walk each kvm mmu page
list under the rcu lock, and we can reschedule as long as we hold a
reference to any mmu page on the list (and unlock the rcu).

For data pages, we call get_page_unless_zero() and verify that it is
the right page, using cmpxchg64 to make sure that the kvm pte hasn't
changed.

BUG=b:269168855
TEST=Ran crostini.VimCompile, borealis Civ VI benchmarks

Change-Id: Ie08fb969f59a2fb854eccae282e283f8ae52597b
Signed-off-by: T.J. Alumbaugh <talumbau@google.com>
Signed-off-by: Yu Zhao <yuzhao@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4246825
Reviewed-by: Yu Zhao <yuzhao@chromium.org>