CHROMIUM: iwl7000: mvm: d3: support v12 wowlan status
Newer firmware versions are going to come with a new version
of the wowlan status API, which gives us replay counters for
both GTKs that might still be in use, as well as unifying the
layout of the counters to just be 64-bit quantities.

Extend our code to handle this case and be able to install
replay counters for both keys, if they're sent up from the
firmware.

Change-Id: I33e27712a12fff2816edc0a8abda1759259e20d7
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 0edd4be915a65e7f171db2383750580fbe294649