CHROMIUM: iwl7000: change HW rev definitions to the new format
The HW REV TYPE values were changed in new HW.  In the past, we used to
convert the new ones into the old format, but that was recently changed.
Now we always use the new format, so the HW revision definitions need to
be updated accordingly.

The way we do that is by using the original values for the new types, so
we don't shift the lower bits.  This means that all values ending with 4
become 1 and all values ending in 8 become 2.

Change-Id: I01abf9c292622b1dafdc92860534fccbce9c3453
Signed-off-by: Mike Golant <michael.golant@intel.com>
iwl7000-tree: 47fb9f423ab4d9593dd29974c3adac5783887370