UPSTREAM: ASoC: soc-pcm: protect BE dailink state changes in trigger
When more than one FE is connected to a BE, e.g. in a mixing use case,
the BE can be triggered multiple times when the FE are opened/started
concurrently. This race condition is problematic in the case of
SoundWire BE dailinks, and this is not desirable in a general
case. The code carefully checks when the BE can be stopped or
hw_free'ed, but the trigger code does not use any mutual exclusion.

Fix by using the same spinlock already used to check FE states, and
set the state before the trigger. In case of errors,  the initial
state will be restored.

This patch does not change how the triggers are handled, it only makes
sure the states are handled in critical sections.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Message-Id: <20210817164054.250028-2-pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 0c75fc7193387776c10f7c7b440d93496e3d5e21)

BUG=b:218766331
TEST=Test sound cards are listed and audio playback works.

Signed-off-by: Terry Cheong <htcheong@chromium.org>
Signed-off-by: Madhulika Dolke <madhulika.dolke@intel.com>
(cherry picked from commit 0c75fc7193387776c10f7c7b440d93496e3d5e21)
Change-Id: Iad29fbf1939ba17b8a037294066143109094b9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3519419
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Sean Paul <seanpaul@chromium.org>