CHROMIUM: iwl7000: fix reset flow for Bz devices
The previous implementation of this was wrong, because the
internal implementation of the CSR_DOORBELL_VECTOR just is
done by connecting it to the UREG_DOORBELL_TO_ISR6. This
means that this was racy, even with the NMI fixed to be at
the right bit position, we could get preempted between the
two writes, and the firmware could process them badly.

Given this understanding of how CSR_DOORBELL_VECTOR works,
and given that the hardware hasn't actually moved the NMI
to BIT(1), at least not yet, fix the reset flow.

Change-Id: I8a224d13a28eb72d4cde2e4920261c400aff7637
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 060709ee159afb0fc39bc0fc29c66efac28b6f1f