CHROMIUM: ASoC: Intel: avs: Support creation and configuration of Peakvol
When peakvol module is being instantiated, use kcontrol state to set
module configuration. To do that, separate module creator is required.
Add it and invoke for peakvol and gain both.

In the past, gain and peakvol have been used simultaneously. The latter
deprecated the former as firmware evolved. For completeness sake, don't
forget about it too.

Signed-off-by: Amadeusz Sławiński <amadeuszx.slawinski@linux.intel.com>

BUG=b:189576468
TEST=none