CHROMIUM: ASoC: Intel: avs: Conditional-path support
Compared to standard paths, conditional paths are not invoked by the
means of FE being opened by userspace and thus their variant selection
is not FE/BE audio format based. These are a sideeffect of standard path
creation if specific criteria are met.

Algorithm is implemented to walk on all existing runtime paths and match
them against conditions provided by topology. If match is found, new
path is created and tied to those which brough it into existence: source
and sink path. If any of the latter perishes, so does the conditional
path.

Conditional paths are used to enable any complex, modern audio scenario
which involve usage of KPB, AEC and WoV modules and more.

Signed-off-by: Cezary Rojewski <cezary.rojewski@intel.com>
Signed-off-by: Amadeusz Sławiński <amadeuszx.slawinski@linux.intel.com>

BUG=b:189576468
TEST=none