UPSTREAM: drm/panel-edp: Split the delay structure out
In the case where we can read an EDID for a panel the only part of the
panel description that can't be found directly from the EDID is the
description of the delays. Let's break the delay structure out so that
we can specify just the delays for panels that are detected by EDID.

This is simple code motion. No functional change is intended.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Sam Ravnborg <sam@ravnborg.org>
Acked-by: Linus Walleij <linus.walleij@linaro.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20210914132020.v5.10.I24f3646dd09954958645cc05c538909f169bf362@changeid
(cherry picked from commit 9ea10a500045767039b2c408738b1c324d7fc4c3)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>
Change-Id: Iece9d54dec3fab700bcdd1dbf481a96f853f0653

BUG=b:228533341, b:220634935
TEST=This patch series can help to pass HDMI 2.1 CTS.

Change-Id: Ib239939dc51da4b2e4cd596f0339dcaeeea9c2db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3743166
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>