UPSTREAM: drm/i915/guc: put all guc objects in lmem when available
The firmware binary has to be loaded from lmem and the recommendation is
to put all other objects in there as well. Note that we don't fall back
to system memory if the allocation in lmem fails because all objects are
allocated during driver load and if we have issues with lmem at that point
something is seriously wrong with the system, so no point in trying to
handle it.

Cc: Matthew Auld <matthew.auld@intel.com>
Cc: Abdiel Janulgue <abdiel.janulgue@linux.intel.com>
Cc: Michal Wajdeczko <michal.wajdeczko@intel.com>
Cc: Vinay Belgaumkar <vinay.belgaumkar@intel.com>
Cc: Radoslaw Szwichtenberg <radoslaw.szwichtenberg@intel.com>
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Matthew Brost <matthew.brost@intel.com>
Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210916162819.27848-3-matthew.brost@intel.com
(cherry picked from commit 7acbbc7cf4851b42b723098aa5699e8fca634e4a)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:240373008
TEST=No gpu hang occur during suspend/resume stress test.

Change-Id: I8c82251cd5f1eb73ef76992b4f90cb26216e5879
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3895960
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>