FROMGIT: iommu/arm-smmu-qcom: Request direct mapping for modem device
The SID configuration requirement for Modem on SC7280 is similar to the
ones found on SC7180/SDM845 SoCs. So, add the SC7280 modem compatible to
the list to defer the programming of the modem SIDs to the kernel.

Signed-off-by: Sibi Sankar <sibis@codeaurora.org>
Reviewed-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/1631886935-14691-5-git-send-email-sibis@codeaurora.org
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit e37f1fe4332491bf2f7b7849d5c3adba0d2a77b3
 https://git.kernel.org/pub/scm/linux/kernel/git/will/linux.git for-joerg/arm-smmu/updates)

BUG=b:177048531
TEST=boot up device

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I244c78b1ec7373a1408c1ed8aa77846c89278263
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3213954
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>