UPSTREAM: drm/i915/tc: Remove waiting for PHY complete during releasing ownership
Waiting for the PHY complete flag to clear when releasing the PHY
ownership was add in

commit ddec362724f9 ("drm/i915: Wait for TypeC PHY complete flag to clear in safe mode")

This isn't required by the spec, the vague idea was to make the
handshake with the firmware more robust, without actual evidence for
when it would be needed. Checking this again, the flag doesn't clear on
ICL until after the PHY's PLL is disabled and the flag is permanently
set on ADL-P. To avoid the spurious timeout messages in dmesg, just
remove this wait.

(cherry-picked from commit 62e1e308ffd7a867ef82375f3cff7f8789ce6721)

BUG=b:216739054
TEST= TypeC DP/HDMI, TBT DP/HDMI

Cc: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210921002313.1132357-4-imre.deak@intel.com
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Change-Id: Iffd8356f57c36d363e477c3ec61100c300d2d8ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3426598
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>