UPSTREAM: drm/i915/tc: Don't keep legacy TypeC ports in connected state w/o a sink
A follow-up patch will disconnect/reconnect PHYs around AUX transfers
and modeset enable/disables. To prepare for that and make things
consistent for all TypeC modes stop connecting the PHY in legacy mode
without a sink being connected. This was done before since in legacy
mode the PHY is dedicated to display usage, so there was no point in
disconnecting it. However after the follow-up changes the TC-cold
blocking power domains will be held as long as the PHY is in the
connected state, so we'll need to disconnect/re-connect the PHY in all
TypeC modes to allow for power saving.

(cherry-picked from commit 675d23c14821fbaef3df5cbad94b42ec2e3a858a)

BUG=b:216739054
TEST=TypeC DP/HDMI, TBT DP/HDMI

Cc: José Roberto de Souza <jose.souza@intel.com>
Signed-off-by: Imre Deak <imre.deak@intel.com>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210921002313.1132357-7-imre.deak@intel.com
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Change-Id: I625d4fb5578410e4224d93a7c5173e881b8b90aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3426601
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>