UPSTREAM: drm/virtio: implement context init: track valid capabilities in a mask
The valid capability IDs are between 1 to 63, and defined in the
virtio gpu spec.  This is used for error checking the subsequent
patches.  We're currently only using 2 capability IDs, so this
should be plenty for the immediate future.

Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>
Acked-by: Lingfeng Yang <lfy@google.com>
Link: http://patchwork.freedesktop.org/patch/msgid/20210921232024.817-4-gurchetansingh@chromium.org
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit 1925d6a7e0f4ce61e35075f87030dcdf512e94dd)

BUG=b:192613935, b:193892617
TEST=glxgears in a crosvm guest.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Change-Id: I44bb5267c0825b742501b89666091805cdeb1f49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3224038
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>