UPSTREAM: drm/virtio: implement context init: handle VIRTGPU_CONTEXT_PARAM_POLL_RINGS_MASK
For the Sommelier guest Wayland proxy, it's desirable for the
DRM fd to be pollable in response to an host compositor event.
This can also be used by the 3D driver to poll events on a CPU
timeline.

This enables the DRM fd associated with a particular 3D context
to be polled independent of KMS events.  The parameter
VIRTGPU_CONTEXT_PARAM_POLL_RINGS_MASK specifies the pollable
rings.

Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>
Acked-by: Nicholas Verne <nverne@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20210921232024.817-11-gurchetansingh@chromium.org
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit 8d6b006e1f51c99016aa39ca9e03947cbdd024e3)

BUG=b:192613935, b:193892617
TEST=glxgears in a crosvm guest.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Change-Id: I739f95c8fed2b017705feafd9275a1ab9f817da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3224205
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>