UPSTREAM: drm/virtio: implement context init: add virtio_gpu_fence_event
Similar to DRM_VMW_EVENT_FENCE_SIGNALED.  Sends a pollable event
to the DRM file descriptor when a fence on a specific ring is
signaled.

One difference is the event is not exposed via the UAPI -- this is
because host responses are on a shared memory buffer of type
BLOB_MEM_GUEST [this is the common way to receive responses with
virtgpu].  As such, there is no context specific read(..)
implementation either -- just a poll(..) implementation.

Signed-off-by: Gurchetan Singh <gurchetansingh@chromium.org>
Acked-by: Nicholas Verne <nverne@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20210921232024.817-12-gurchetansingh@chromium.org
Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
(cherry picked from commit cd7f5ca33585918febe5e2f6dc090a21cfa775b0)

BUG=b:192613935, b:193892617
TEST=glxgears in a crosvm guest.

Signed-off-by: Ryan Neph <ryanneph@google.com>
Change-Id: I159324cc0c0759d403a8b244bd058e5c3c4e72ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3224206
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>