UPSTREAM: drm/i915/gem: Implement a function to process all gem objects of a region
An upcoming common pattern is to traverse the region object list and
perform certain actions on all objects in a region. It's a little tricky
to get the list locking right, in particular since a gem object may
change region unless it's pinned or the object lock is held.

Define a function that does this for us and that takes an argument that
defines the action to be performed on each object.

v3:
- Improve structure documentation a bit (Matthew Auld)

Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>
Reviewed-by: Matthew Auld <matthew.auld@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20210922062527.865433-3-thomas.hellstrom@linux.intel.com
(cherry picked from commit d80ee88e0769e2e05afeb5d04b4dc43fc107b0d5)
Signed-off-by: Lee Shawn C <shawn.c.lee@intel.com>

BUG=b:240373008
TEST=No gpu hang occur during suspend/resume stress test.

Change-Id: Ice45348a342c03bf850f427123ee51ec3e4893a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3895966
Tested-by: Shawn C Lee <shawn.c.lee@intel.com>
Commit-Queue: Shawn C Lee <shawn.c.lee@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>