CHROMIUM: clk: mediatek: make pwrmcu clock as always on
The clocks for pwrmcu is used by sspm module which is a crucial part in
suspend/resume flow. However, there is no corresponding pwrmcu driver in
kernel, and this situation will let clock driver to disable the clocks
for pwrmcu and let suspend/resume flow failed.

We could add the pwrmcu driver for chromium project. However, there is no
other requirements for pwrmcu in chromium project, the pwrmcu driver
is only added to keep the clock on. For this case, we can just make the main
clocks of pwrmcu as always on to solve it.

BUG=b:190783437
TEST=emerge-cherry sys-kernel/chromeos-kernel-5_10

Signed-off-by: Chun-Jie Chen <chun-jie.chen@mediatek.com>
Change-Id: I2e07da02e89c0b5eeb5230c880ab98f0dd6eb386
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3177801
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>