FROMLIST: iommu/mediatek: Initialise/Remove for multi bank dev
The registers for each bank of the IOMMU base are in order, delta is
0x1000. Initialise the base for each bank.

For all the previous SoC, we only have bank0. thus use "do {} while()"
to allow bank0 always go.

When removing the device, Not always all the banks are initialised, it
depend on if there is masters for that bank.

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
(am from https://patchwork.kernel.org/project/linux-mediatek/patch/20210923115840.17813-32-yong.wu@mediatek.com/)

BUG=b:188609090
TEST=Boot OK

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: I34d43b92f6a8a86dc257d6173dfebbe292f29699
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3215701
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Tested-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>