UPSTREAM: ath11k: Fix the spectral minimum FFT bin count
User was not able to configure the spectral with the FFT bin count 32.
In all supported platforms, the expected minimum FFT bin count is 32 but
it was wrongly defined as 64. This restrict the user to not configure
down to the actually supported minimum FFT bin count. So update the
minimum FFT bin count as 32.

Tested-on: IPQ8074 hw2.0 AHB WLAN.HK.2.4.0.1-01492-QCAHKSWPL_SILICONZ-1
Tested-on: IPQ6018 hw1.0 AHB WLAN.HK.2.4.0.1-00330-QCAHKSWPL_SILICONZ-1

Signed-off-by: Karthikeyan Periyasamy <periyasa@codeaurora.org>
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20210721180809.90960-4-jouni@codeaurora.org
(cherry picked from commit 6dfd20c8a6cd1fcf2c68d86c9d678f42535f6ade)

BUG=b:205920109
TEST=wifi_matfunc/perf

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I4f37c7dc0a6d11f464a1e17eca1c7acd769c5298
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3351366
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>