UPSTREAM: Bluetooth: Fix handling of experimental feature for codec offload
The existence of the experimental feature identifiy is the indication
that it is supported or not. No extra flag needed and the initial flag
should define if a feature is enabled or not. This is actually defined
in the management API definition.

Fixes: ad93315183285 ("Bluetooth: Add offload feature under experimental flag")
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
(cherry picked from commit 7f7fd17ed7c5228229f55f2082c0fd784cc4dad4)

BUG=b:183594508
TEST=Build

Signed-off-by: Yong Zhi <yong.zhi@intel.com>
Change-Id: I9edb179229561e613df6d5810635d4af5cd75b60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3365575
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Sathya Prakash M R <sathya.prakash.m.r@intel.com>
Commit-Queue: Michael Sun <michaelfsun@google.com>