BACKPORT: FROMGIT: arm64: dts: qcom: pmk8350: Make RTC disabled by default; enable on sc7280-idp
The RTC on the pmk8350 is not useful on all boards. Some boards may
not provide backup power to the PMIC but might have another RTC on the
board that does have backup power. In this case it's better to not use
the RTC on the PMIC.

At the moment, the only boards that includes this PMIC are sc7280-idp
and sc7280-idp2. On sc7280-idp I'm not aware of any other RTCs, but
sc7280-idp2 has a Chrome OS EC on it and this is intended to provide
the RTC for the AP.

Let's do what we normally do for hardware that's not used by all
boards and set it to a default status of "disabled" and then enable it
on the boards that need it.

NOTE: for sc7280-idp it's _possible_ we might also want to add
`allow-set-time;`. That could be the subject of a future patch if it
is indeed true.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Satya Priya <skakit@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
[bjorn: Enable the RTC on the MTP as well]
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20210929153553.1.Ib44c2ac967833d7a3f51452d44d15b7b8d23c1f0@changeid

(cherry picked from commit 12a7f71a8ea5218e004f7226a7cb0d1005952c85
 https://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git arm64-for-5.16)

Conflicts:
   arch/arm64/boot/dts/qcom/sm8350-mtp.dts

Drop sm8350-mtp part because we don't want to pull in a bunch of patches
that aren't relevant.

BUG=b:186387242
TEST=emerge-herobrine chromeos-kernel-5_10; lspci

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ib44c2ac967833d7a3f51452d44d15b7b8d23c1f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3252403
Reviewed-by: Douglas Anderson <dianders@chromium.org>