UPSTREAM: drm/i915: Remove memory frequency calculation
This memory frequency calculated is only used to check if it is zero,
what is not useful as it will never actually be zero.

Also the calculation is wrong, we should be checking other bit to
select the appropriate frequency multiplier while this code is stuck
with a fixed multiplier.

So here dropping it as whole.

v2:
- Also remove memory frequency calculation for gen9 LP platforms

Cc: Yakui Zhao <yakui.zhao@intel.com>
Cc: Matt Roper <matthew.d.roper@intel.com>
Fixes: 5d0c938ec9cc ("drm/i915/gen11+: Only load DRAM information from pcode")
Signed-off-by: José Roberto de Souza <jose.souza@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211013010046.91858-1-jose.souza@intel.com
(cherry picked from commit 83f52364b15265aec47d07e02b0fbf4093ab8554)

BUG=b:228272527
BUG=b:179453336
TEST=for arch in amd64-generic arm64-generic; do \
	USE=buildtest emerge-"${arch}" chromeos-kernel-5_10; \
     done

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I3c77d8bed03ddc7911f5674da107a9e849cdd6ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3573806
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>